# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VETRA API - Platform Kesehatan Hewan Peliharaan",
    "description": "Dokumentasi Lengkap API untuk VETRA - Platform Booking Klinik, Konsultasi Dokter, Chatting, dan AI Chatbot",
    "version": "1.0.0",
    "contact": {
      "name": "Vetra API Support",
      "email": "support@vetra.com"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/user/profile": {
      "get": {
        "tags": ["User Profile"],
        "summary": "Mengambil detail profil pengguna",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "Berhasil mengambil profil" }
        }
      },
      "put": {
        "tags": ["User Profile"],
        "summary": "Memperbarui profil pengguna",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "User Updated" },
                  "phone": { "type": "string", "example": "081234567890" },
                  "address": { "type": "string", "example": "Jl. Kamboja No. 9" },
                  "password": { "type": "string", "format": "password" },
                  "password_confirmation": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profil berhasil diperbarui" }
        }
      }
    },
...
    "/user/pets": {
      "get": {
        "tags": ["Pets"],
        "summary": "Get list hewan peliharaan user",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "Daftar hewan peliharaan" }
        }
      },
      "post": {
        "tags": ["Pets"],
        "summary": "Tambah hewan peliharaan baru",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "species"],
                "properties": {
                  "name": { "type": "string", "example": "Milo" },
                  "species": { "type": "string", "example": "cat" },
                  "breed": { "type": "string", "example": "Anggora" },
                  "age": { "type": "integer", "example": 2 },
                  "weight": { "type": "number", "example": 3.4 },
                  "gender": { "type": "string", "enum": ["male", "female"], "example": "male" },
                  "color": { "type": "string", "example": "Orange" },
                  "medical_history": { "type": "string", "example": "Vaksin Tahunan" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Hewan berhasil ditambahkan" }
        }
      }
    },
    "/user/pets/{id}": {
      "put": {
        "tags": ["Pets"],
        "summary": "Update data hewan",
        "security": [{"bearerAuth": []}],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "age": { "type": "integer" },
                  "weight": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Data hewan berhasil diupdate" }
        }
      },
      "delete": {
        "tags": ["Pets"],
        "summary": "Hapus data hewan",
        "security": [{"bearerAuth": []}],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Hewan berhasil dihapus" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Autentikasi standar JWT. Masukkan token Anda (Tanpa menulis 'Bearer ')"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Autentikasi integrasi API Key statis. Kirimkan token API Key Anda pada header X-API-Key"
      }
    },
...
  }
}